nouveau reducer pour eviter dempiler le cra
