+++ src/actions/index.js
@@ -3,3 +3,4 @@ export * from './conseiller.actions';
 export * from './structure.actions';
 export * from './cra.actions';
 export * from './menu.actions';
+export * from './statistique.actions';
