+++ src/components/connected/statistics/Components/ElementDatePicker.js
@@ -1,6 +1,6 @@
 import React, { forwardRef, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { craActions } from '../../../../actions';
+import { statistiqueActions } from '../../../../actions';
 import PropTypes from 'prop-types';
 import DatePicker, { registerLocale } from 'react-datepicker';
 import fr from 'date-fns/locale/fr';
@@ -8,19 +8,19 @@ import fr from 'date-fns/locale/fr';
 registerLocale('fr', fr);
 function ElementDatePicker(props) {
 
-  let dateDebut = useSelector(state => state.cra?.dateDebutStats);
-  let dateFin = useSelector(state => state.cra?.dateFinStats);
+  let dateDebut = useSelector(state => state.statistique?.dateDebutStats);
+  let dateFin = useSelector(state => state.statistique?.dateFinStats);
 
   const dispatch = useDispatch();
   const setDate = date => {
     if (props.nomDate === 'datePickerDebut') {
       if (date <= dateFin) { //on ne change la date de debut que si elle est <= date de fin
-        dispatch(craActions.changeDateStatsDebut(date));
-        dispatch(craActions.getStatsCra(date, dateFin));
+        dispatch(statistiqueActions.changeDateStatsDebut(date));
+        dispatch(statistiqueActions.getStatsCra(date, dateFin));
       }
     } else if (date >= dateDebut) { //on ne change la date de fin que si elle est >= date de début
-      dispatch(craActions.changeDateStatsFin(date));
-      dispatch(craActions.getStatsCra(dateDebut, date));
+      dispatch(statistiqueActions.changeDateStatsFin(date));
+      dispatch(statistiqueActions.getStatsCra(dateDebut, date));
     }
   };
